feature/core functions

+++ app/Http/Controllers/API/ProjectController.php
@@ -15,10 +15,19 @@ class ProjectController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    
+
     public function index()
     {
-        $Projects = Project::all();
-        return response(['Projects' => ProjectResource::collection($Projects)]);
+        $pageSize = isset($_GET['pageSize'])?$_GET['pageSize']:2;
+        $sortBy = isset($_GET['sortBy'])?$_GET['sortBy']:'name';
+        $sortDirection = isset($_GET['sortDirection'])?$_GET['sortDirection']:'ASC';
+        $Projects = Project::orderBy($sortBy, $sortDirection);
+        $q = isset($_GET['q'])?$_GET['q']:'';
+        $Projects = $q?$Projects->where('name', 'LIKE', "%$q%"):$Projects;
+        $Projects = $Projects->paginate($pageSize);
+        return response(['Projects' => $Projects]);
+        
     }
 
     /**
@@ -27,8 +36,20 @@ class ProjectController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
+
+    private function checkUserRole()
+    {
+        if(auth('api')->user()->role != 'PRODUCT_OWNER'){
+            return false;
+        }
+        return true;
+    }
+
     public function store(Request $request)
     {
+        if(!$this->checkUserRole()){
+            return Response::deny('You must be a PRODUCT_OWNER to continue.');
+        }
         $data = $request->all();
 
         $validator = Validator::make($data, [
@@ -52,6 +73,7 @@ class ProjectController extends Controller
      */
     public function show(Project $Project)
     {
+        
         return response(['Project' => new ProjectResource($Project)]);
     }
 
@@ -64,6 +86,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, Project $Project)
     {
+        if(!$this->checkUserRole()){
+            return Response::deny('You must be a PRODUCT_OWNER to continue.');
+        }
         $data = $request->all();
 
         $validator = Validator::make($data, [
@@ -87,6 +112,9 @@ class ProjectController extends Controller
      */
     public function destroy(Project $Project)
     {
+        if(!$this->checkUserRole()){
+            return Response::deny('You must be a PRODUCT_OWNER to continue.');
+        }
         $Project->delete();
 
         return response(['message' => 'Project deleted successfully']);
